+++ src/main/todo-list/src/route/Todo.jsx
@@ -1,114 +1,43 @@
-import { useEffect, useState } from "react";
-import styles from "./css/Todo.module.css";
-import Input from "./css/component/Input";
+import {useState} from "react";
 
-const Todo = () => {
-  const [todo, setTodo] = useState("");
-  const [deadline, setDeadlineDate] = useState("");
-  const [estimatedTime, setEstimatedTime] = useState(0);
-  const [difficulty, setDifficulty] = useState("mid");
-  const [todos, setTodos] = useState([]);
+import styles from "./css/Todo.module.css"
+import Input from "./css/component/Input"
+import TodoInput from "./css/component/TodoInput";
+import Importance from "./css/component/Importance";
 
-  useEffect(() => {
-    const currentDate = new Date();
-    const dateStr = currentDate.toLocaleDateString('en-CA');
-    setDeadlineDate(dateStr);
-  }, [])
+const Test = () => {
+    const [todoContent, setTodoContent] = useState("")
 
-  const handlers = {
-    onChangeTodo: (e) => setTodo(e.target.value),
-    onChangeDeadlineDate: (e) => setDeadlineDate(e.target.value),
-    onClickDifficulty: (e) => setDifficulty(e.target.value),
-    onChangeEstimatedTime: (e) => setEstimatedTime(e.target.value),
-    onClickSubmit: (e) => {
-      e.preventDefault();
-      setTodos((prevTodos) => [
-        ...prevTodos,
-        {
-          todo: todo,
-          date: deadline,
-          time: estimatedTime,
-          difficulty: difficulty,
-        },
-      ]);
-      setTodo("");
-    },
-  };
-
-  return (
-    <div className={styles.todo_list}>
-      <form>
-          <Input
-            type = "text"
-            label = "할 일"
-            value = {todo}
-            id = "todo"
-            onChange = {handlers.onChangeTodo}
-          /> <br />
-          
-          <Input
-            type = "date"
-            label = "마감기한"
-            value = {deadline}
-            id = "deadline"
-            onChange = {handlers.onChangeTodo}
-          /> <br />
-          
-          <Input
-            type = "number"
-            label = "예상소요시간"
-            value = {estimatedTime}
-            id = "estimatedTime"
-            onChange = {handlers.onChangeEstimatedTime}
-          /> <br />
-
-          <h3>난이도</h3>
-          <Input
-            type = "radio"
-            label = "high"
-            name = "difficulty"
-            id = "high"
-            onChange = {handlers.onClickDifficulty}
-          />
-
-          <Input
-            type = "radio"
-            label = "mid"
-            name = "difficulty"
-            id = "mid"
-            onChange = {handlers.onClickDifficulty}
-            DC = {true}
-
-          />
-
-          <Input
-            type = "radio"
-            label = "low"
-            name = "difficulty"
-            id = "low"
-            onChange = {handlers.onClickDifficulty}
-          /> <br />
-
-
-        <button
-          onClick={handlers.onClickSubmit}
-        >
-          완료
-        </button>
-      </form>
-      {todos.map((td, idx) => {
-        return (
-          <div key={idx}>
-            <span>{idx + 1}</span>||
-            <span>{td.todo}</span>||
-            <span>{td.date}</span>||
-            <span>{td.time}</span>||
-            <span>{td.difficulty}</span>
-          </div>
-        );
-      })}
-    </div>
-  );
+    return (
+        <div className={styles.test}>
+            <div>
+                <div>
+                </div>
+            </div>
+            <div className={styles.inputs}>
+                <Input />
+                <div>
+                    <TodoInput 
+                    iconName="hotel_class"
+                    description="중요도"
+                    Component={<Importance />}
+                    />
+                    <TodoInput 
+                    iconName="event"
+                    description="기한"
+                    />
+                    <TodoInput 
+                    iconName="timer"
+                    description="예상 소요 시간"
+                    />
+                    <TodoInput 
+                    iconName="device_thermostat"
+                    description="난이도"
+                    />
+                </div>
+            </div>
+        </div>
+    );
 }
 
-export default Todo;
+export default Test;
