--- src/main/todo-list/src/route/css/Test.module.css
@@ -1,44 +1,0 @@
-.test {
-    width: 100%;
-
-    display: flex;
-    /* height: 100vh; */
-    flex-direction: column;
-    /* flex-grow: 1; */
-}
-
-.test div:nth-child(1) {
-    flex-grow: 9;
-}
-
-
-.test div:nth-child(2) {
-    flex-grow: 1;
-}
-
-.inputs > span {
-    height: 2rem;
-    width: 100%;
-    text-align: center;
-}
-
-.inputs span input {
-    height: 100%;
-    width: 90%;
-}
-
-.inputs > div {
-    display: flex;
-    flex-direction: row;
-    justify-content: flex-end;
-    width: 90%;
-}
-
-.inputs > div span{
-    width: 50px;
-    font-size: 2rem;
-}
-
-.inputs > div span:hover{
-    cursor: pointer;
-}
